refact: create user component and upload image components
