+++ src/app/features/users/presentation/components/create-update-user/create-update-user.component.html
@@ -1,0 +1,89 @@
+<div class="card w-full shadow-md border border-gray-200">
+        <form class="flex flex-row" [formGroup]="createUpdateUserFormServices.createOrUpdateUserForm" autocomplete="off">
+            <div class="flex-1 w-full">
+
+                <div class="card-body"> <h1 class="text-2xl font-bold"> Creación de usuario </h1> </div>
+        
+                <div class="flex flex-col gap-y-8 px-6 pb-8">
+
+                    <app-generic-input
+                        label="Nombre(s)"
+                        id="firstName"
+                        controlName="firstName"
+                        [formGroup]="createUpdateUserFormServices.createOrUpdateUserForm">
+                    </app-generic-input>
+
+                    <app-generic-input
+                        label="Apellidos"
+                        id="lastName"
+                        controlName="lastName"
+                        [formGroup]="createUpdateUserFormServices.createOrUpdateUserForm">
+                    </app-generic-input>
+
+                    <app-generic-input
+                        label="Correo"
+                        id="email"
+                        controlName="email"
+                        [formGroup]="createUpdateUserFormServices.createOrUpdateUserForm">
+                    </app-generic-input>
+
+                    @if ( createUpdateUserFormServices.isUpdateUser() ) {
+                        <div class="flex flex-col gap-4">
+                            <label class="font-semibold">Estatus</label>
+                            <label class="flex items-center gap-3 cursor-pointer w-12">
+                            <input 
+                                type="checkbox"
+                                formControlName="active"
+                                class="toggle toggle-primary w-12 border rounded-2xl"
+                                [checked]="createUpdateUserFormServices.isActive"
+                                (change)="($event)"/>
+                            <span>
+                                @if ( createUpdateUserFormServices.isActive ) {
+                                    Activo
+                                } @else {
+                                    Inactivo
+                                }
+                            </span>
+                            </label>
+                        </div>
+                    }
+
+                    <app-password-input
+                        label="Contraseña"
+                        id="password"
+                        controlName="password"
+                        [formGroup]="createUpdateUserFormServices.createOrUpdateUserForm">
+                    </app-password-input>
+                </div>
+            </div>
+
+            <div class="p-20 flex flex-col gap-y-8 flex-1 items-start">
+
+                <input 
+                    type="file" 
+                    accept="image/png,image/jpeg" 
+                    class="file-input file-input-lg w-2/3" 
+                    #fileInput 
+                    (change)="uploadImageManagment.onFileSelected($event)"/>
+
+                @if ( uploadImageManagment.imagePreview() ) {
+                    <div class="flex flex-col gap-y-8 items-center">
+                        
+                        @if( uploadImageManagment.imagePreview().includes('blob') ) {
+                            <img [src]="uploadImageManagment.imagePreview()" alt="Preview" class="w-60 h-60 rounded-full border-none"/>
+                        } @else {
+                            <div class="avatar">
+                                <div class="w-56">
+                                    <img ngSrc="{{ uploadImageManagment.imagePreview() }}" fill alt="Log image" class="rounded-full border-none" priority/>
+                                </div>
+                            </div>
+                        }
+    
+                        <button type="button" class="btn btn-error text-white btn-sm w-fit" (click)="uploadImageManagment.removeImage()">
+                            Eliminar imagen
+                        </button>
+                    </div>
+                }
+            </div>
+        </form>
+    </div>
