+++ src/app/features/users/presentation/components/create-update-user/index.ts
@@ -1,0 +1,1 @@
+export * from './create-update-user.component';
