+++ src/app/features/users/presentation/components/index.ts
@@ -1,2 +1,4 @@
 export * from './filters-selects';
+export * from './roles-list-select';
 export * from './delete-user-dialog';
+export * from './create-update-user';
