+++ src/app/features/users/presentation/components/roles-list-select/roles-list-select.component.css
@@ -1,0 +1,3 @@
+:host {
+  display: block;
+}
