+++ src/app/features/users/presentation/components/roles-list-select/roles-list-select.component.html
@@ -1,0 +1,28 @@
+<div class="flex flex-row gap-x-8 w-2/5">
+
+    @if ( rolesList.isLoading() ) {
+        <div class="card flex-1 shadow-md p-8 border border-gray-200 justify-center items-center">
+            <span class="loading loading-ring loading-xl"></span>
+        </div>
+    } @else {
+        <div class="card flex-1 shadow-md p-8 border border-gray-200">
+            <div class="flex gap-x-4">
+                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" class="w-8 text-blue-700" fill="currentColor">
+                    <path d="M3.78307 2.82598L12 1L20.2169 2.82598C20.6745 2.92766 21 3.33347 21 3.80217V13.7889C21 15.795 19.9974 17.6684 18.3282 18.7812L12 23L5.6718 18.7812C4.00261 17.6684 3 15.795 3 13.7889V3.80217C3 3.33347 3.32553 2.92766 3.78307 2.82598ZM5 4.60434V13.7889C5 15.1263 5.6684 16.3752 6.7812 17.1171L12 20.5963L17.2188 17.1171C18.3316 16.3752 19 15.1263 19 13.7889V4.60434L12 3.04879L5 4.60434ZM12 11C10.6193 11 9.5 9.88071 9.5 8.5C9.5 7.11929 10.6193 6 12 6C13.3807 6 14.5 7.11929 14.5 8.5C14.5 9.88071 13.3807 11 12 11ZM7.52746 16C7.77619 13.75 9.68372 12 12 12C14.3163 12 16.2238 13.75 16.4725 16H7.52746Z"></path>
+                </svg>
+                <p class="text-2xl"> Selección de Roles </p>
+            </div>
+            @for (item of rolesList.data()?.data; track item.idRoles) {
+                <label class="label mt-6 w-1/4">
+                    <input 
+                        type="checkbox" 
+                        class="checkbox checkbox-primary" 
+                        [checked]="createUpdateUserFormServices.isRolChecked().includes(item.idRoles)" 
+                        (change)="createUpdateUserFormServices.onRoleToggle(item.idRoles, $event.target.checked)"/>
+                        {{ item.name }}
+                </label>
+            }
+        </div>
+    }
+
+</div>
