+++ src/app/features/users/presentation/services/index.ts
@@ -1,2 +1,4 @@
+export * from './create-update-user-form.service';
 export * from './select-filters-users.service';
+export * from './uploadImageManagmente.service';
 export * from './user-selected.service';
